Optimize day 23 challenge 2
No need to use an actual node object to represent the linked list.
Because the value space is the same as the cardinality and the values
are numbers that can be used as array indices, an array suffices as a
linked list where every value is the next index in the linked list.